refactor: some refactor

+++ src/resources/tasks/task.router.js
@@ -1,4 +1,5 @@
 const router = require('express').Router({ mergeParams: true });
+const { StatusCodes, getReasonPhrase } = require('http-status-codes');
 const Task = require('./task.model');
 const tasksService = require('./tasks.service');
 
@@ -8,11 +9,11 @@ router.route('/').get(async (req, res) => {
   } = req;
   const tasks = await tasksService.getAll(boardId);
   if (Array.isArray(tasks)) {
-    res.status(200);
+    res.status(StatusCodes.OK);
     res.json(tasks.map(Task.toResponse));
   } else {
-    res.status(404);
-    res.json({ message: "There aren't tasks with such boardId" });
+    res.status(StatusCodes.NOT_FOUND);
+    res.json({ message: getReasonPhrase(StatusCodes.NOT_FOUND) });
   }
 });
 
@@ -20,13 +21,18 @@ router.route('/:taskId').get(async (req, res) => {
   const {
     params: { boardId, taskId },
   } = req;
-  const task = await tasksService.getTaskById(boardId, taskId);
-  if (!task) {
-    res.status(404);
-    res.json({ message: "Task with such parameters isn't find" });
+  if (taskId) {
+    const task = await tasksService.getTaskById(boardId, taskId);
+    if (!task) {
+      res.status(StatusCodes.NOT_FOUND);
+      res.json({ message: getReasonPhrase(StatusCodes.NOT_FOUND) });
+    }
+    res.status(StatusCodes.OK);
+    res.json(task);
+  } else {
+    res.status(StatusCodes.BAD_REQUEST);
+    res.json({ message: getReasonPhrase(StatusCodes.BAD_REQUEST) });
   }
-  res.status(200);
-  res.json(task);
 });
 
 router.route('/').post(async (req, res) => {
@@ -34,13 +40,18 @@ router.route('/').post(async (req, res) => {
     body,
     params: { boardId },
   } = req;
-  try {
-    const task = await tasksService.createTask({ ...body, boardId });
-    res.status(201);
-    res.json(task);
-  } catch {
-    res.status(500);
-    res.json({ message: 'Something went wrong' });
+  if (boardId) {
+    try {
+      const task = await tasksService.createTask({ ...body, boardId });
+      res.status(StatusCodes.CREATED);
+      res.json(task);
+    } catch {
+      res.status(StatusCodes.INTERNAL_SERVER_ERROR);
+      res.json({ message: getReasonPhrase(StatusCodes.INTERNAL_SERVER_ERROR) });
+    }
+  } else {
+    res.status(StatusCodes.BAD_REQUEST);
+    res.json({ message: getReasonPhrase(StatusCodes.BAD_REQUEST) });
   }
 });
 
@@ -48,17 +59,18 @@ router.route('/:taskId').delete(async (req, res) => {
   const {
     params: { boardId, taskId },
   } = req;
-  try {
-    const result = await tasksService.deleteTasks(boardId, [taskId]);
-    if (result === null) {
-      res.status(400);
-      res.json({ message: 'Bad request' });
+  if (boardId && taskId) {
+    try {
+      await tasksService.deleteTasks(boardId, [taskId]);
+      res.status(StatusCodes.NO_CONTENT);
+      res.json({ boardId, taskId });
+    } catch {
+      res.status(StatusCodes.INTERNAL_SERVER_ERROR);
+      res.json({ message: getReasonPhrase(StatusCodes.INTERNAL_SERVER_ERROR) });
     }
-    res.status(204);
-    res.json({ boardId, taskId });
-  } catch {
-    res.status(500);
-    res.json({ message: 'Something went wrong' });
+  } else {
+    res.status(StatusCodes.BAD_REQUEST);
+    res.json({ message: getReasonPhrase(StatusCodes.BAD_REQUEST) });
   }
 });
 
@@ -67,13 +79,18 @@ router.route('/:taskId').put(async (req, res) => {
     params: { boardId, taskId },
     body,
   } = req;
-  try {
-    await tasksService.updateTask(boardId, taskId, body);
-    res.status(200);
-    res.json({ message: 'Task is updated successfully' });
-  } catch (err) {
-    res.status(500);
-    res.json({ message: 'Something went wrong' });
+  if (boardId && taskId) {
+    try {
+      await tasksService.updateTask(boardId, taskId, body);
+      res.status(StatusCodes.OK);
+      res.json({ message: 'Task is updated successfully' });
+    } catch (err) {
+      res.status(StatusCodes.INTERNAL_SERVER_ERROR);
+      res.json({ message: getReasonPhrase(StatusCodes.INTERNAL_SERVER_ERROR) });
+    }
+  } else {
+    res.status(StatusCodes.BAD_REQUEST);
+    res.json({ message: getReasonPhrase(StatusCodes.BAD_REQUEST) });
   }
 });
