+++ src/resources/users/user.router.js
@@ -1,11 +1,11 @@
 const router = require('express').Router();
+const { StatusCodes, getReasonPhrase } = require('http-status-codes');
 const User = require('./user.model');
 const usersService = require('./user.service');
 
 router.route('/').get(async (req, res) => {
   const users = await usersService.getAll();
-  // map user fields to exclude secret fields like "password"
-  res.status(200);
+  res.status(StatusCodes.OK);
   res.json(users.map(User.toResponse));
 });
 
@@ -13,30 +13,34 @@ router.route('/:userId').get(async (req, res) => {
   const {
     params: { userId },
   } = req;
-  const user = await usersService.getUserById(userId);
-  // map user fields to exclude secret fields like "password"
-  if (user) {
-    res.status(200);
-    res.json(user);
+  if (userId) {
+    const user = await usersService.getUserById(userId);
+    if (user) {
+      res.status(StatusCodes.OK);
+      res.json(User.toResponse(user));
+    } else {
+      res.status(StatusCodes.NOT_FOUND);
+      res.json({ message: getReasonPhrase(StatusCodes.NOT_FOUND) });
+    }
   } else {
-    res.status(404);
-    res.json({ message: 'User not found' });
+    res.status(StatusCodes.BAD_REQUEST);
+    res.json({ message: getReasonPhrase(StatusCodes.BAD_REQUEST) });
   }
 });
 
 router.route('/').post(async (req, res) => {
   const { body } = req;
   if (!body.name || !body.login || !body.password) {
-    res.status(400);
-    res.json({ message: 'Bad request' });
+    res.status(StatusCodes.BAD_REQUEST);
+    res.json({ message: getReasonPhrase(StatusCodes.BAD_REQUEST) });
   }
   try {
     const user = await usersService.createUser(body);
-    res.status(201);
-    res.json(user);
+    res.status(StatusCodes.CREATED);
+    res.json(User.toResponse(user));
   } catch (err) {
-    res.status(500);
-    res.json({ message: 'Something went wrong' });
+    res.status(StatusCodes.INTERNAL_SERVER_ERROR);
+    res.json({ message: getReasonPhrase(StatusCodes.INTERNAL_SERVER_ERROR) });
   }
 });
 
@@ -45,13 +49,18 @@ router.route('/:userId').put(async (req, res) => {
     params: { userId },
     body,
   } = req;
-  try {
-    await usersService.updateUser(userId, body);
-    res.status(200);
-    res.json({ message: 'User is updated successfully' });
-  } catch (err) {
-    res.status(500);
-    res.json({ message: 'Something went wrong' });
+  if (userId && body.name && body.login && body.password) {
+    try {
+      const user = await usersService.updateUser(userId, body);
+      res.status(StatusCodes.OK);
+      res.json(user);
+    } catch (err) {
+      res.status(StatusCodes.INTERNAL_SERVER_ERROR);
+      res.json({ message: getReasonPhrase(StatusCodes.INTERNAL_SERVER_ERROR) });
+    }
+  } else {
+    res.status(StatusCodes.BAD_REQUEST);
+    res.json({ message: getReasonPhrase(StatusCodes.BAD_REQUEST) });
   }
 });
 
@@ -59,9 +68,14 @@ router.route('/:userId').delete(async (req, res) => {
   const {
     params: { userId },
   } = req;
-  await usersService.deleteUser(userId);
-  res.status(204);
-  res.json({ userId });
+  if (userId) {
+    await usersService.deleteUser(userId);
+    res.status(StatusCodes.NO_CONTENT);
+    res.json({ userId });
+  } else {
+    res.status(StatusCodes.BAD_REQUEST);
+    res.json({ message: getReasonPhrase(StatusCodes.BAD_REQUEST) });
+  }
 });
 
 module.exports = router;
